Follow MS-SMB2 for error handling
In order to parse a response as an ErrorResponse, require that the
structure size is equal to 9. From MS-SMB2 2.2.2:
  StructureSize (2 bytes): The server MUST set this field to 9

This allows for IOCTLs to return an error status without pike
attempting to parse an ErrorResponse structure from the response.

For example in 3.3.5.15.6.2:
The server MUST construct an SMB2 IOCTL Response, following the syntax
specified in section 2.2.32, with the following values:

    Status in the SMB2 header MUST be set to STATUS_INVALID_PARAMETER.