handle EALREADY internally in the transport
When EALREADY is raised during a write, call a new method on the poller,
defer_write, which begins monitoring the transport (connection) for
write events. When a write event is received, handle_write is called
once from the poller to allow the application to retry the write.

Normally sockets are not monitored for write events as they are
generally always writable and the client application decides when to
send writes. In most cases, read events are the only thing
the application cares about handling from the event loop. Now the event
loop will handle the retry logic.

This also adds docstrings to all functions in the transport module.

Minor fix in model.py to reraise the error with the traceback for easier
debugging.

Bump version number to 0.2.1