implement pre-auth integrity hash per session
MS-SMB2 3.2.5.3.1
If Connection.Dialect is "3.1.1", the client MUST compute its
preauthentication integrity hash value as follows:

    Set Session.PreauthIntegrityHashValue to
    Connection.PreauthIntegrityHashValue.

    The client MUST generate a hash using the
    Connection.PreauthIntegrityHashId algorithm on the string
    constructed by concatenating Session.PreauthIntegrityHashValue and
    the session setup request message retrieved from the
    Connection.OutstandingRequests. The client MUST set
    Session.PreauthIntegrityHashValue to the hash value generated above.

    The client MUST generate a hash using the
    Connection.PreauthIntegrityHashId algorithm on the string
    constructed by concatenating Session.PreauthIntegrityHashValue and
    the session setup response message, including all bytes from the
    response's SMB2 header to the last byte received from the network.
    The client MUST set Session.PreauthIntegrityHashValue to the hash
    value generated above.

The previous implementation was deficient in that it calculated the
session PreAuthIntegrityHashValue in the context of the connection which
was only correct for the first session established on a connection.
Subsequent sessions would base their Session.PreauthIntegrityHashValue
on the previous session's hash value as opposed to the
Connection.PreauthIntegrityHashValue and ultimately derive the wrong
signing keys.

A test has been added to pike/test/session.py to multiplex 3 sessions on
a single connection and assert that they share a connection while
deriving different session keys.

A nice side-benefit of this refactoring has been localization of the
hash calculation to the areas near negotiate and session_setup as
opposed to a method on the Connection which was called for all packets
in the connection stream. This may provide a slight performance benefit.