ErrorResponse: never decode the "extra" byte, it is ignored anyway
Smb2.decode will handle advancing the cursor to the next command or end
of buffer, so there is no point in decoding the extra byte as specified
in [MS-SMB2] 2.2.2. Additionally since some versions of Windows no
longer write the "extra" byte when error_count and byte_count is zero,
this can lead to BufferOverflow exception.