Raise RequestError when attempting to submit non-Netbios frames
Previously, pass an object that isn't a Netbios frame to submit or
transceive will cause the pike internal loop to hang -- additionally no
Future is allocated so the 30-second default timeout does not apply.

This change allows pike to fail fast when an inappropriate object is
supplied. This is mainly for the benefit of the programmer during test
development.