test/changnotify: hit STATUS_NOTIFY_CLEANUP path
When the closing a handle with a pending notify request, the server responds
with STATUS_NOTIFY_CLEANUP. This exposes a bug in pike where the OutputBufferOffset
field of the notify response is interpreted as the ErrorContextCount
field of an error response. If the value is non-zero, pike will attempt
to read the error context header off the end of the packet.