pike.model: query_file_info behavioral pre-deprecation
Add new parameter `first_result_only` to `query_file_info` that
currently defaults to None.

When first_result_only is True, the previous behavior is retained and the
function returned the first child of the QueryInfoResponse

When first_result_only is False, the QueryInfoResponse itself is returned.

When the value is None, a DeprecationWarning is emitted indicating that
the default behavior will change in a future release. The current
default behavior is first_result_only=True.