Vendor py3c library
For allow building of cross-platform wheels, bundle
the py3c library with pike so that it doesn't
need to be installed.

On CentOS 6 / 7 (manylinux builder images), the
py3c package only exists for x86_64.

On alpine (musllinux builder images), there is no
packaged py3c, so there is no way to install it.